Base logging to file added.
 - Base logging implemented. It writes logs with datetime, filename (in which log was raised), level and message

 - Echo commands while creating db set to False

 - repr method for representing a hero a bit changed, to be one-line

 - not-nullable constraints to db added

+++ repositories.py
@@ -1,9 +1,10 @@
+import logging
 import random
 from datetime import datetime
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.sql import functions, Insert
+from sqlalchemy.sql import functions
 
 from models import Hero, Sides, Motto, Story, Battle
 
@@ -11,6 +12,24 @@ engine = create_engine('postgresql+psycopg2://postgres_user:rootroot@localhost:5
 Session = sessionmaker(bind=engine)
 session = Session()
 
+# Требования к logging:
+#  - минимум 1 фильтр на события (например, не писать лог если бой закончился вничью).
+
+#  - один хендлер с отправкой части логов в консоль.
+#  - один хендлер с отправкой части логов в другой файл.
+#
+#  - НЕОБЯЗАТЕЛЬНО: хендлер с отправкой error & critical логов в телеграмм бота с помощью telegram api.
+#    Бот сразу пишет сообщение либо в группу, id которой указан в env, либо в личку пользователю, id которого указан в env. Я должен быть добавлен в эту группу: @rioran
+
+logging.basicConfig(
+    format="%(asctime)s => %(filename)s => %(levelname)s => %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+    level=logging.INFO
+)
+
+#01:24:09
+console = logging.StreamHandler()
+console.setLevel(logging.ERROR)
 
 def add_hero(name: str,
              birthday: datetime,
@@ -21,6 +40,8 @@ def add_hero(name: str,
     new_hero = Hero(name=name, birthday=birthday, tribe=tribe, side=side, power=power)
     session.add(new_hero)
     session.commit()
+
+    logging.info(f'Hero was created: {new_hero}')
 
 
 class HeroNotFoundException(Exception):
@@ -36,7 +57,9 @@ def add_motto(hero_id: int, motto: str) -> None:
         hero.mottos.append(Motto(motto_id=motto_id, motto=motto))
 
         session.commit()
+        logging.info(f'Motto {motto} for hero {hero_id} was added')
     else:
+        logging.critical(f'Hero not found')
         raise HeroNotFoundException
 
 
@@ -47,19 +70,27 @@ def add_story(hero_id: int, story: str) -> None:
     if hero:
         if hero.story:
             hero.story.story = story
+            logging.info(f'story {story} for hero {hero_id} was updated')
         else:
             hero.story = Story(story=story)
+            logging.info(f'Story {story} for hero {hero_id} was added')
 
         session.commit()
     else:
+        logging.critical(f'Hero not found')
         raise HeroNotFoundException
 
 
-def delete_hero(hero_id) -> None:
+def delete_hero(hero_id: int) -> None:
     """Deletes hero from table. With hero deletes users story and all mottos"""
     hero = session.query(Hero).get(hero_id)
-    session.delete(hero)
-    session.commit()
+
+    if hero:
+        session.delete(hero)
+        session.commit()
+        logging.info(f'Hero {hero_id} successfully deleted')
+    else:
+        logging.warning(f'Hero {hero_id} doesn\'t exist - probably as you wanted')
 
 
 def add_battle() -> None:
@@ -80,3 +111,4 @@ def add_battle() -> None:
 
     session.add(new_battle)
     session.commit()
+    logging.info(f'Battle between {first_warrior.name} and {second_warrior.name} happened. Winner: {winner}')
